finish read data

+++ src/com/example/btdemo/communication/data/DeviceInfo.java
@@ -3,6 +3,6 @@ package com.example.btdemo.communication.data;
 public class DeviceInfo {
 	public String serialNumber;
 	public byte[] btaddress;
-	public int version;
+	public long version;
 	public String deviceName;
 }
